readonlyInStagingMode default to false (#26004)
This pull request makes a small change to the default data store
policies in the `dataStoreRuntime.ts` file. The change updates the
default setting for `readonlyInStagingMode` to be `false` instead of
`true`, allowing data stores to be writable in staging mode by default.
This is to aid prototyping. StagingMode is still alpha, so this breaking
change is acceptable.