Add API to erase schema details, and use it on table (#25999)
## Description

Adds experimental alpha API for doing component style implementation
detail erasure of schema.

These are then used on the table component.

Interestingly the new APIs do not show up in the types of code which
uses them, so exporting the new APIs was not required to use them on
table, but they are exported anyway since we want to make them available
as alpha. Them being alpha should not block table from going to beta.

## Breaking Changes

Some implementation details of table have been hidden.