Use new strictEnum instead of branding for SchemaFormatVersion (#26012)
## Description

The current code does a poor job of type safety (the members of the
constant are not branded, so the branding forces lots of unsafe type
conversions and is likely less type safe than not branding would have
been).

Four solutions to this were considered:
1. Remove the branding: this would remove all the calls to brand, which
are inferring what they are branding to and thus remove the risk of them
silently doing an unintended type conversion since they simply convert
to whatever is needed at that location. This would improve the type
safety of all other branded types by reducing the risky use of branding.
2. Brand the constant values: this would centralize the branding to
where its clearly correct and easy to review, improving type safety.
This could be done with brandConst, but there is an annoyance that the
current branded type is a union of the two values, so we can't brand to
that easily, and would want to introduce a second type, one for the
branded values, and one for the union of the branded constants.
3. Implement a custom utility to help do option 2 above.
4. An enum: this gets some of the type safety benefits of branding, but
it allows number to be assigned to it, and if there are string values
allows them to be assigned without branding.
5. const enums: same as above, but instead of a regression in code size,
this is an improvement. Const enums can cause issues if you both package
export them, and change the values used. As we don't intend to ever
package export this and can't change the values (as they are used in
persisted data), this seems safe for two redundant reasons.

I think the best options are either const enums or option 3, a custom
enum utility.

I have gone with option 3 due to the surprising lack of safety with the
build in enums with respect to `number`.