Make tree's "currentVersion" actually current. (#26033)
## Description

"currentVersion" Inside of shared tree is currently unused except for
tests.

It was set to 2.0.0, and this makes it actually stary current.

As this is only used in tests, all the snapshot tests with update simply
reflect a change to how their tests are configured, ensuring they test
for the latest version and can detect changes in it.