fix(map): Emit valueChanged event when there are pending local ops after remote clear (#26010)
When a remote clear operation is received while there are pending local
set operations, the map emits a "clear" event but the map still contains
keys. This made the clear event unreliable as an indicator of map state,
requiring the oracle to re-read all keys after every clear event.

This change adds functionality to emit `valueChanged` event for keys
which are deleted during remote clears. This way the customers would
have to just focus on the `valueChanged` event and not `clear` event for
remote ops.


[ADO#48665](https://dev.azure.com/fluidframework/internal/_workitems/edit/48665)

Follow-up:
[ADO#55457](https://dev.azure.com/fluidframework/internal/_workitems/edit/55457)