Make squash parameter required on internal plumbing (#26081)
## Description

Make the `squash` parameter required on a number of functions only
currently intended to be called internally (but which are exposed in the
legacy API surface). The optionality of squashing on the main entry
point (`commitStagedChanges`) has been preserved. See #25762 for more
details.

## Breaking Changes

Direct callers of this function can pass `false` where they previously
passed no argument.
The breaking change here is largely only at compile time, as the
falsiness of `undefined` ends up resulting in equivalent behavior at the
usage site in Fluid code.

---------

Co-authored-by: Abram Sanderson <absander@microsoft.com>