Fix result folder path to work for vscode and npm terminal runs (#26092)
## Description

In tree-agent-langchain, when running integration tests the results are
placed in different paths as the code currently doesn't handle having
different working directories. This change figures out the the dirname
using the new added file and adds the dir `integration-test-results`
directory so that the test results can be written there. (Had to add
`../../src/test/integration-test-results` because the tests actually get
run in lib/test not src/test.) Note that this dir was always created
whenever integration tests were run but was never tracked by git and
should continue to not be.

## Reviewer Guidance

* Originally tried to do it by relative path using the current file's
location. Tried using `import.meta.url` but that wasn't defined in
CommonJS. Tried directly using `__dirname` but that isn't defined in ESM
* Followed other tests that needed this and added a new `dirname.cjs`
file that exports the dirname properly.